transition dark brem process to a G4VDiscreteProcess
dark brem model is now ldmx-specific and not inheriting from Geant4
nonsense

dark brem configuration and process allow for different models to be
written in the future

dark brem process can be biased using our normal biasing framework
