+++ python/simcfg.py
@@ -89,63 +89,3 @@ class PrimaryGenerator:
         """
 
         return '%s of class %s' % (self.instance_name, self.class_name)
-
-class DarkBrem:
-    """Storage for parameters of dark brem process
-
-    Attributes
-    ----------
-    APrimeMass : float
-        Mass of A' in MeV
-    library_path : str
-        Path to directory of LHE files containing dark brem vertices
-    enable : bool
-        Should we use the custom Geant4 dark brem process?
-    method : int
-        Interpretation method for LHE files
-    threshold : float
-        Minimum energy that electron should have for dark brem to have nonzero xsec
-    epsilon : float
-        Epsilon for dark brem xsec calculation
-    only_one_per_event : bool
-        Should we deactivate the process after one dark brem or allow for more than one?
-    """
-
-    def __init__(self) : 
-
-        self.APrimeMass   = 0.
-        self.library_path = ''
-        self.enable       = False #off by default
-        self.method       = 1
-        self.threshold    = 0.
-        self.epsilon      = 1.
-        self.only_one_per_event = False
-
-    def activate(self, apMass, lib_path ) :
-        """Activate the dark brem process with the input A' mass [MeV] and path to vertex library"""
-
-        self.enable       = True
-        self.APrimeMass   = apMass
-        self.library_path = lib_path
-
-    def __str__(self): 
-        """Stringify the DarkBrem configuration
-
-        Returns
-        -------
-        str
-            A human-readable version of all its attributes
-        """
-
-        string  = "{ Enabled: %r"%self.enable
-        if self.enable :
-            string += ", Mass: %.1f MeV"%self.APrimeMass
-            string += ", Library: '%s'"%self.library_path
-            string += ", Method: %d"%self.method
-            string += ", Threshold: %.2f GeV"%self.threshold
-            string += ", Epsilon: %.2E"%self.epsilon
-            string += ", Only One Per Event: %r"%self.only_one_per_event
-
-        string += " }"
-
-        return string
